added decimal functionability

+++ Program.cs
@@ -5,7 +5,8 @@
 // int c = a + b;
 // Console.WriteLine(c);
 
-int num1 = 0; int num2 = 0;
+float num1 = 0;
+float num2 = 0;
 
 Console.WriteLine("Console Calculator in C#\r");
 Console.WriteLine("------------------------\n");
